+++ app/src/main/java/rectangledbmi/com/pittsburghrealtimetracker/patterns/response/PatternResponse.java
@@ -1,5 +1,5 @@
 
-package rectangledbmi.com.pittsburghrealtimetracker.model;
+package rectangledbmi.com.pittsburghrealtimetracker.patterns.response;
 
 import javax.annotation.Generated;
 import com.google.gson.annotations.Expose;
